(feat): create single product
-make a get  request for a single product.

[(Finishes #161296536)]
