+++ server/db.js
@@ -1,11 +1,15 @@
-const db = [
-  {
-    id: 1,
-    name: 'macbook',
-    desc: 'good laptop',
-    stock: 2,
-    availableStock: 1,
-  },
-];
+const db = {
+  products: [
+    {
+      id: 1,
+      name: 'macbook',
+      desc: 'good laptop',
+      stock: 2,
+      availableStock: 1,
+    },
+  ],
+  sales: [],
+};
+
 
 export default db;
